fix: Dynamic @test definitions do not work
As explained [here][1]() and [here][2], the `@test` definition is pre-processed by `bats` and replaced with a single function definition, so attempting to dynamically create test cases in a loop does not work.

There is little value gained using the current approach anyway, so instead switching to a more traditional approach of defining each test case explicitly and having the logic to run the test in a function called from each test, to avoid duplication of logic.

[1]: https://stackoverflow.com/questions/24443777/how-do-you-programmatically-add-a-bats-test
[2]: https://github.com/bats-core/bats-core/wiki/Bats-Evaluation-Process